Generate maps only with Wall borders
For the Automata generator, we make sure that the border tiles start as
Walls, since after that each iteration do not change them already.

For the Drunkard generator, we break the iteration when we arrive at the
border. Potentially we could adjust the RNG to never go into a border
instead, this way each drunk walk iteration would live for longer but this
would make the logic way more complicated and I'm not sure it would
yield better results.

Fixes #6