Check if dest position is still available
Because our move logic runs in two steps, first emitting all the move
intentions and then processing those, it can happen that some of them
clash when 2 entities wants to move to the same position.

So we need to keep track of already occupied spaces in the map and
just skip our move if the space became occupied since the intention to
move was created.

Fixes #20